Spot requests never expire
This was causing spot instance requests to expire and on subsequent
deploys of a service tf would think it needed to make another request
and instance even though an instance still existed and it was only the
request that was gone.

This will keep the spot instance request around forever (unless it can
no longer be fulfilled for what I assume would be pricing, etc reasons).