Update spot instance volume size when changed
It seems like there is a bug in terraform for the
aws_spot_instance_request. When updating the volume size and doing a
`terraform plan` it shows that it will be updating it, but it doesn't.

This script looks up the attached volume and and updates it with the
specified volume size

Note that this script assumes that there is only 1 attached volume.